Changed implementation to i128 to support larger numbers

+++ src/lenstra.rs
@@ -1,13 +1,13 @@
 use num_integer::gcd;
 use rand::RngExt;
 
-pub fn ecm(n: i64) -> (i64, i64) {
+pub fn ecm(n: i128) -> (i128, i128) {
     let mut rng = rand::rng();
 
     loop {
-        let cap_a: i64 = rng.random::<i64>() % n;
-        let a: i64 = rng.random::<i64>() % n;
-        let b: i64 = rng.random::<i64>() % n;
+        let cap_a: i128 = rng.random::<i128>() % n;
+        let a: i128 = rng.random::<i128>() % n;
+        let b: i128 = rng.random::<i128>() % n;
 
         let cap_p = (a, b);
         // let cap_b = (b * b - a * a * a - cap_a * a) % n;
@@ -15,16 +15,16 @@ pub fn ecm(n: i64) -> (i64, i64) {
         // E: y^2 = (x^3 + Ax + B) mod n
 
         let mut iter = 2;
-        let mut d: i64 = 0;
+        let mut d: i128 = 0;
         let mut div_failed = false;
         let mut prev_point = cap_p;
 
         while !div_failed {
             let mut temp_point = prev_point;
-            let mut dividend: i64;
-            let mut divisor: i64;
+            let mut dividend: i128;
+            let mut divisor: i128;
             for _ in 1..iter {
-                let alpha: i64;
+                let alpha: i128;
                 if prev_point == temp_point {
                     dividend = 3 * temp_point.0 * temp_point.0 + cap_a;
                     divisor = 2 * temp_point.1;
@@ -53,7 +53,7 @@ pub fn ecm(n: i64) -> (i64, i64) {
 }
 
 // based on pseudocode from wikipedia
-fn mod_inverse(a: i64, n: i64) -> i64 {
+fn mod_inverse(a: i128, n: i128) -> i128 {
     let mut t = 0;
     let mut newt = 1;
     let mut r = n;
