+++ src/main.rs
@@ -11,7 +11,7 @@ fn main() {
     io::stdin()
         .read_line(&mut input)
         .expect("Failed to read line");
-    let n: i64 = input.trim().parse().expect("Please type a valid integer");
+    let n: i128 = input.trim().parse().expect("Please type a valid integer");
 
     println!("Factoring: {}", n);
     let result = ecm(n);
